--- src/components/tabs/tabs.jsx
@@ -1,43 +1,0 @@
-import "./tabs.css";
-import { engine } from "../../engine";
-import { Component } from 'react';
-import { buildTab } from "@coveo/headless";
-
-class Tabs extends Component {
-
-  constructor(props) {
-    super(props);
-
-    const initTab = {
-      options: { expression: props.tabExpression },
-      initialState: { isActive: props.defaultTab }
-    };
-    this.headlessTab = buildTab(engine, initTab);
-    this.state = this.headlessTab.state;
-  }
-
-  componentDidMount() {
-    this.headlessTab.subscribe(() => this.setState(this.headlessTab.state));
-  }
-
-  render() {
-    return (
-      <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand" href="#">Navbar</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
-          <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
-          <div class="navbar-nav">
-            <a class="nav-item nav-link active" href="#">Home <span class="sr-only">(current)</span></a>
-            <a class="nav-item nav-link" href="#">Features</a>
-            <a class="nav-item nav-link" href="#">Pricing</a>
-            <a class="nav-item nav-link disabled" href="#">Disabled</a>
-          </div>
-        </div>
-      </nav>
-    );
-  }
-}
-
-export default Tabs;
