+++ src/index.js
@@ -5,4 +5,4 @@ import App from './App';
 import 'bootstrap/dist/css/bootstrap.css'
 import 'react-bootstrap-typeahead/css/Typeahead.css';
 
-ReactDOM.render(<App/>, document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById("root"));
